+++ resources/views/project/setting-basic.blade.php
@@ -41,7 +41,7 @@
                 </div>
                 <div class="mb-3">
                     <label for="project-sort" class="bmd-label-floating">项目排序（值越大越靠后）</label>
-                    <input type="number" name="sort_level" class="form-control float-left w-75" id="project-sort" value="{{ old('sort_level', $project->sort_level) }}" {{ Auth::user()->can('project-sort') ? '' : 'disabled' }}/>
+                    <input type="number" name="sort_level" class="form-control w-75" id="project-sort" value="{{ old('sort_level', $project->sort_level) }}" {{ Auth::user()->can('project-sort') ? '' : 'disabled' }}/>
                     <i class="fa fa-question-circle ml-2" data-bs-toggle="tooltip" title="" data-original-title="只有管理员可以修改"></i>
                 </div>
                 <div class="mb-3">