+++ resources/views/search.blade.php
@@ -24,7 +24,7 @@
                 @endif
             </div>
             <div class="card-header-operation">
-                <div class="bmd-form-group bmd-collapse-inline pull-right">
+                <div class="pull-right">
                     @if (empty($tag) && !Auth::guest())
                     <span class="mr-2 text-warning">∞</span>
                     <select class="wz-search-range-change wz-select-card-title text-warning">
@@ -41,13 +41,11 @@
                 <form id="wz-search-box" action="{{ wzRoute('search:search') }}" method="get">
                     <div class="row marketing wz-main-container-full search-panel">
                         <div class="input-group mb-3">
-                            <input type="text" class="form-control mr-3" placeholder="输入要搜索的文档标题" name="keyword"
+                            <input type="text" class="form-control" placeholder="输入要搜索的文档标题" name="keyword"
                                    value="{{ $keyword ?? '' }}">
                             <input type="hidden" name="project_id" value="{{ $project_id ?? '' }}">
                             <input type="hidden" name="range" value="{{ $range ?? '' }}">
-                            <div class="input-group-append">
-                                <button class="btn btn-outline-primary" type="submit">搜索</button>
-                            </div>
+                            <button class="btn btn-outline-primary" type="submit">搜索</button>
                         </div>
                     </div>
                 </form>