+++ resources/views/user/basic.blade.php
@@ -24,7 +24,7 @@
                            name="username">
                 </div>
                 <div class="mb-3">
-                    <div class="col-sm-offset-2 col-sm-10">
+                    <div>
                         <button type="submit"
                                 class="btn btn-success btn-raised">@lang('common.btn_save')</button>
                     </div>